Small fixes after PR. Change `maxPartSize` type to `int`, fix buffer initialization, and correct Polish typo in logging message.

+++ ksef-client/src/main/java/io/alapierre/ksef/batch/model/BatchConfig.java
@@ -8,7 +8,7 @@ import java.nio.file.Path;
  */
 public record BatchConfig(
         Path outputDir,
-        long maxPartSize,
+        int maxPartSize,
         boolean cleanup
 ) {
 }
